backend/requirements.txt
Flask
Flask-Cors
gunicorn
numpy
torch
facenet-pytorch
scikit-learn
psycopg2-binary
Pillow